backend/requirements/base.txt
# Core Django
Django>=4.2,<5.0
psycopg2-binary>=2.9.9

# Django REST Framework
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Image handling
Pillow>=10.0.0

# Environment variables
python-decouple>=3.8
